+++ Sources/DTModelStorage/SectionModel.swift
@@ -28,6 +28,14 @@ import Foundation
 /// Data holder for single section in `MemoryStorage`.
 open class SectionModel: Section, SupplementaryAccessible
 {
+    
+    /// Returns item at index, if it exists, nil otherwise.
+    /// - Parameter index: Index to search for.
+    open func item(at index: Int) -> Any? {
+        guard items.count > index else { return nil }
+        return items[index]
+    }
+    
     /// Items for current section
     /// - Warning: If you try to set new array to this property [T], the only way to do this without exception is to wrap it into items.map { $0 }. This is a workaround that exists because of Swift inability to cast [T] to [Any]. You can call `setItems` method instead of doing so.
     /// - SeeAlso: `setItems:`